+++ src/main/java/dev/latvian/mods/kubejs/item/custom/SwordItemBuilder.java
@@ -14,7 +14,7 @@ public class SwordItemBuilder extends HandheldItemBuilder {
 
 	public SwordItemBuilder(ResourceLocation i) {
 		super(i, 3F, -2.4F);
-		itemAttributeModifiers = SwordItem.createAttributes(toolTier, attackDamageBaseline, speedBaseline);
+		itemAttributeModifiers = SwordItem.createAttributes(toolTier, (int) attackDamageBaseline, speedBaseline);
 		parentModel = SWORD_MODEL;
 		tag(SWORD_TAGS);
 	}
